Improving more imports and moving everything for aggregation into own module
